requirements.txt
django
djang-ninja
requests
python-dotenv